First commit that actually everything connects together and works
Renamed dir gvirtus-role-detector to deprecated-gvirtus-role-detector as it is no longer used. Created the gvirtus backend daemonset and service. Fixed logic on when the gvirtus device plugin is applied based on node labels. Renamed dir 'app' to 'gvirtus-example-app'. Add dir 'gvirtus-labeler' which basically replaces the deprecated gvirtus-role-detector. It automatically adds label in all cluster nodes that show if a node can be used as frontend or backend. Renamed dir gvirtus-webhook to gvirtus-mutating-webhook. Deleted gvirtus-webhook/server.py.old